Create a separate header for data.
